backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
neo4j
psycopg2-binary
requests
camelot-py[cv]
